KVM: arm64: Invoke the snapshot interface for the host stage-2 pagetable
Allocate memory for the snapshot by creating a memory cache with empty
pages that will be used by the hypervisor during the page table copy.
Get the required size of the PGD and allocate physically contiguous
memory for it. Allocate contiguous memory for an array that is used to
keep track of the pages used from the memcache. Call the snapshot
interface and release the memory for the snapshot.

Signed-off-by: Sebastian Ene <sebastianene@google.com>
Message-Id: <20231218135859.2513568-5-sebastianene@google.com>