KVM: arm64: Walk the pagetable snapshot and parse the ptdump descriptors
Define the attributes for the stage-2 pagetables that will be used by
the ptdump parser code to interepret and show the descriptors. Build the
pagetable level description dynamically and use the KVM pagetable walker
to visit the PTEs. Display the number of the bits used by the IPA
address space and the start level.

Signed-off-by: Sebastian Ene <sebastianene@google.com>
Message-Id: <20231218135859.2513568-9-sebastianene@google.com>