irqchip/gic-v4.1: Fix skipping VMOVP in cpu offline scenario
commit dd3f050a216e ("irqchip/gic-v4.1: Implement the v4.1 flavour
of VMOVP") make an optimization, VMOVP can be skipped if moving
VPE to a cpu whose RD is sharing its VPE table with the current one.

In the cpu offline scenario, the mask_val is the entire cpu range,
except for the offline cpu. Therefore, the first cpu is CPU0. In
corner case, this may result in lost interrupts:
0. Each cpu die shares a VPE table and contains 32 CPUs
   die0(CPU0-31) die1(CPU32-63)...
1. VPE resides on CPU32, doorbell affinity to CPU32.
2. Move VPE to CPU16, doorbell affinity to CPU16.
3. Manually offline CPU16 on the host side:
   'echo 0 > /sys/devices/system/cpu/cpu16/online'
4. VMOVP will be skipped.
5. Subsequent doorbell interrupts will be lost.

So VMOVP cannot be skipped when the affinity CPU is not in mask_val.

Fixes: dd3f050a216e ("irqchip/gic-v4.1: Implement the v4.1 flavour of VMOVP")
Signed-off-by: Kunkun Jiang <jiangkunkun@huawei.com>
Message-Id: <20240126103012.1020-3-jiangkunkun@huawei.com>