KVM: arm64: Override Microsoft Azure Cobalt 100 MIDR value with ARM Neoverse N2
Several workload optimizations and errata depend on validating that the
optimization or errata are applicable to the particular CPU by checking
the MIDR_EL1 system register value. With the Microsoft implementer ID
for Azure Cobalt 100, the value doesn't match and ~20-25% performance
regression is seen in these workloads. Override the Azure Cobalt 100
value and replace it with the default ARM Neoverse N2 value that Azure
Cobalt 100 is based on.

Signed-off-by: Easwar Hariharan <eahariha@linux.microsoft.com>
Message-Id: <20240206195819.1146693-1-eahariha@linux.microsoft.com>