KVM: arm64: Add tracepoints + stats for LPI cache effectiveness
LPI translation and injection has been shown to have a significant
impact on the performance of VM workloads, so it probably makes sense to
add some signals in this area.

Introduce the concept of a KVM tracepoint that associates with a VM
stat and use it for the LPI translation cache tracepoints. It isn't too
uncommon for a kernel hacker to attach to tracepoints, while at the same
time userspace may open a 'binary stats' FD to peek at the corresponding
VM stats.

Signed-off-by: Oliver Upton <oliver.upton@linux.dev>
Message-Id: <20240213093250.3960069-2-oliver.upton@linux.dev>