KVM: arm64: Add early_param to control WFx trapping
Add an early_param to control WFx (WFI or WFE) trapping. This is so
interrupts can be passed through if the CPU has support for direct
interrupt injection, a feature of GICv4. This is described as an
enumeration with three possible behaviors, always passthrough (never
trap), never passthrough (always trap), or default (trap if more than
one task is running. Default matches the current behavior.

Signed-off-by: Colton Lewis <coltonlewis@google.com>
Message-Id: <20240410175437.793508-1-coltonlewis@google.com>