KVM: arm64: Store the maximum sve vector length at hyp
In subsequent patches, hyp needs to know the maximum sve vector
length for the host, without needing the trust the host for that.
This is used when allocating memory for the host sve state in the
following patch, as well as for allocating and restricting guest
sve state in a future patch series.

Signed-off-by: Fuad Tabba <tabba@google.com>
Message-Id: <20240517131814.719933-5-tabba@google.com>