KVM: arm64: Relax locking for kvm_test_age_gfn and kvm_age_gfn
Replace the MMU write locks for read locks.

Grabbing the read lock instead of the write lock is safe because the
only requirement we have is that the stage-2 page tables do not get
deallocated while we are walking them. The stage2_age_walker() callback
is safe to race with itself; update the comment to reflect the
synchronization change.

Signed-off-by: James Houghton <jthoughton@google.com>
Message-Id: <20240529180510.2295118-7-jthoughton@google.com>