arm64: Override set_fixmap_io
Override the set_fixmap_io to set shared permission for the host
in case of a CC guest. For now we mark it shared unconditionally.

If/when support for device assignment and device emulation in the realm
is added in the future then this will need to filter the physical
address and make the decision accordingly.

Signed-off-by: Suzuki K Poulose <suzuki.poulose@arm.com>
Signed-off-by: Steven Price <steven.price@arm.com>
Tested-by: Itaru Kitayama <itaru.kitayama@fujitsu.com>
Message-Id: <20240605093006.145492-7-steven.price@arm.com>