KVM: x86: Implement kvm_fast_test_age_gfn and kvm_fast_age_gfn
The fast-only notifier will only report an accurate result when the
shadow MMU is not in use.

Implement kvm_arch_young_notifier_likely_fast(), as MGLRU will check
this function to see if it should even be attempting the fast-only
notifier. We only want to attempt the notifier if there is a chance that
it will succeed (i.e., that we're using the TDP MMU).

Signed-off-by: James Houghton <jthoughton@google.com>
Message-Id: <20240611002145.2078921-8-jthoughton@google.com>