arm64: Deprecate old local_daif_{mask,save,restore} helper functions
Motivation
----------

With upcoming FEAT_NMI extension in Arm64 v8.8, the kernel needs to
support two separate interrupt masking implementation. This increase
complexity in the exception masks management, furthermore, the FEAT_NMI
feature introduces the ALLINT feild in PSTATE, rendering the current
helper functions inadequate to reflect the actual behavior.

This patch deprecates the old interrupt masking helper functions
starting with local_daif_. Developers are encouraged to migrate to the
new series of logical interrupt masking function starting with
local_allint_.

Implementation
--------------

This patch replaces the instance of the old functions with their
corresponding new couterparts:

- local_daif_mask() -> local_allint_mask()

- local_daif_save_flags() -> local_allint_save_flags()

- local_daif_restore(flags) -> local_allint_restore(flags), except for
  specific cases. It always used with local_allint_save_flags() in pair.

- local_daif_restore(DAIF_PROCCTX_NOIRQ) -> local_nmi_serror_enable()

- local_daif_restore(DAIF_ERRCTX) -> local_nmi_serror_disable()

- local_daif_restore(DAIF_PROCCTX) -> local_irq_serror_enable()

Benefits
--------

The new API functions offer clear naming that reflect their purpose,
regardless of the kernel NMI configuration. This provides developers
with a consistent and comprehensive set of function to use, even when
the FEAT_NMI feature is enabled in the future.

Signed-off-by: Liao Chang <liaochang1@huawei.com>
Message-Id: <20240614034433.602622-7-liaochang1@huawei.com>